setting individual Units
