+++ extras/AFC_BoxTurtle.py
@@ -1,7 +1,38 @@
+# Armored Turtle Automated Filament Changer
+#
+# Copyright (C) 2024 Armored Turtle
+#
+# This file may be distributed under the terms of the GNU GPLv3 license.
+
+
 class afcBoxTurtle:
     def __init__(self, config):
         self.printer = config.get_printer()
+        self.AFC = self.printer.lookup_object('AFC')
         self.printer.register_event_handler("klippy:connect", self.handle_connect)
+        self.name = config.get_name().split()[-1]
+        self.type='Box Turtle'
+        self.screen_mac = config.get('screen_mac', None)
+        self.lanes=[]
+        self.AFC.units[self.name]=config.get_name().split()[0]
+ 
+        self.led_name =config.get('led_name',self.AFC.led_name)
+        self.led_fault =config.get('led_fault',self.AFC.led_fault)
+        self.led_ready = config.get('led_ready',self.AFC.led_ready)
+        self.led_not_ready = config.get('led_not_ready',self.AFC.led_not_ready)
+        self.led_loading = config.get('led_loading',self.AFC.led_loading)
+        self.led_prep_loaded = config.get('led_loading',self.AFC.led_prep_loaded)
+        self.led_unloading = config.get('led_unloading',self.AFC.led_unloading)
+        self.led_tool_loaded = config.get('led_tool_loaded',self.AFC.led_tool_loaded)
+
+    def get_status(self, eventtime=None):
+        self.response = {}
+        self.response['name'] = self.name
+        self.response['type'] = 'Box Turtle'
+        self.response['screen'] = self.screen_mac
+        self.response['lanes'] = self.lanes
+        
+        return self.response
 
     def handle_connect(self):
         """
@@ -9,8 +40,6 @@ class afcBoxTurtle:
         This function is called when the printer connects. It looks up AFC info
         and assigns it to the instance variable `self.AFC`.
         """
-        self.AFC = self.printer.lookup_object('AFC')
-
         firstLeg = '<span class=warning--text>|</span><span class=error--text>_</span>'
         secondLeg = firstLeg + '<span class=warning--text>|</span>'
         self.logo ='<span class=success--text>R  _____     ____\n'
@@ -25,8 +54,6 @@ class afcBoxTurtle:
         self.logo_error+='O |              \ \n'
         self.logo_error+='R |          |\ <span class=secondary--text>X</span> |\n'
         self.logo_error+='! \_________/ |___|</span>\n'
-
-        self.AFC.gcode.register_mux_command('CALIBRATE_AFC', None, None, self.cmd_CALIBRATE_AFC, desc=self.cmd_CALIBRATE_AFC_help)
 
     def system_Test(self, UNIT, LANE, delay, assignTcmd):
         msg = ''
@@ -100,179 +127,7 @@ class afcBoxTurtle:
 
         return succeeded
 
-    cmd_CALIBRATE_AFC_help = 'calibrate the dist hub for lane and then afc_bowden_length'
-    def cmd_CALIBRATE_AFC(self, gcmd):
-        """
-        This function performs the calibration of the hub and Bowden length for one or more lanes within an AFC
-        (Automated Filament Changer) system. The function uses precise movements to adjust the positions of the
-        steppers, check the state of the hubs and tools, and calculate distances for calibration based on the
-        user-provided input. If no specific lane is provided, the function defaults to notifying the user that no lane has been selected. The function also includes
-        the option to calibrate the Bowden length for a particular lane, if specified.
+    
 
-        Usage:`CALIBRATE_AFC LANES=<lane> DISTANCE=<distance> TOLERANCE=<tolerance> BOWDEN=<lane>`
-        Examples:
-            - `CALIBRATE_AFC LANES=all Bowden=leg1 DISTANCE=30 TOLERANCE=3`
-            - `CALIBRATE_AFC BOWDEN=leg1` (Calibrates the Bowden length for 'leg1')
-
-        Args:
-            gcmd: The G-code command object containing the parameters for the command.
-                Parameters:
-                - LANES: Specifies the lane to calibrate. If not provided, calibrates no lanes.
-                - DISTANCE: The distance to move during calibration (optional, defaults to 25mm).
-                - TOLERANCE: The tolerance for fine adjustments during calibration (optional, defaults to 5mm).
-                - BOWDEN: Specifies the lane to perform Bowden length calibration (optional).
-
-        Returns:
-            None
-        """
-        dis = gcmd.get_float('DISTANCE', 25)
-        tol = gcmd.get_float('TOLERANCE', 5)
-        afc_bl = gcmd.get('BOWDEN', None)
-        short_dis = self.AFC.short_move_dis
-        lanes = gcmd.get('LANES', None)
-
-        if self.AFC.current is not None:
-            self.AFC.gcode.respond_info('Tool must be unloaded to calibrate Bowden length')
-            return
-
-        cal_msg = ''
-
-        def find_lane_to_calibrate(lane_name):
-            """
-            Search for the given lane across all units in the AFC system.
-
-            Args: lane_name: The name of the lane to search for
-
-            Returns: The lane name if found, otherwise None
-            """
-            for UNIT in self.AFC.unit.keys():
-                if lane_name in self.AFC.unit[UNIT]:
-                    return lane_name
-
-            # If the lane was not found
-            self.AFC.gcode.respond_info('{} not found in any unit.'.format(lane_name))
-            return None
-
-        # Helper functions for movement and calibration
-        def calibrate_hub(CUR_LANE, CUR_HUB):
-            hub_pos = 0
-            hub_pos = move_until_state(CUR_LANE, lambda: CUR_HUB.state, CUR_HUB.move_dis, tol, short_dis, hub_pos)
-            tuned_hub_pos = calc_position(CUR_LANE, lambda: CUR_HUB.state, hub_pos, short_dis, tol)
-            return tuned_hub_pos
-
-        def move_until_state(lane, state, move_dis, tolerance, short_move, pos=0):
-            while state() == False:
-                lane.move(move_dis, self.AFC.short_moves_speed, self.AFC.short_moves_accel)
-                pos += move_dis
-            self.AFC.reactor.pause(self.AFC.reactor.monotonic() + 0.1)
-            while state() == True:
-                lane.move(short_move * -1, self.AFC.short_moves_speed, self.AFC.short_moves_accel, True)
-                pos -= short_move
-            self.AFC.reactor.pause(self.AFC.reactor.monotonic() + 0.1)
-            while state() == False:
-                lane.move(tolerance, self.AFC.short_moves_speed, self.AFC.short_moves_accel)
-                pos += tolerance
-            return pos
-
-        def calc_position(lane, state, pos, short_move, tolerance):
-            while state():
-                lane.move(short_move * -1, self.AFC.short_moves_speed, self.AFC.short_moves_accel, True)
-                pos -= short_move
-            self.AFC.reactor.pause(self.AFC.reactor.monotonic() + 0.1)
-            while not state():
-                lane.move(tolerance, self.AFC.short_moves_speed, self.AFC.short_moves_accel)
-                pos += tolerance
-            return pos
-
-        def calibrate_lane(LANE):
-            if LANE not in self.AFC.stepper:
-                self.AFC.gcode.respond_info('{} Unknown'.format(LANE.upper()))
-                return
-            CUR_LANE = self.AFC.stepper[LANE.name]
-            CUR_HUB = self.printer.lookup_object('AFC_hub {}'.format(CUR_LANE.unit))
-            if CUR_HUB.state:
-                self.AFC.gcode.respond_info('Hub is not clear, check before calibration')
-                return False, ""
-            if not CUR_LANE.load_state:
-                self.AFC.gcode.respond_info('{} not loaded, load before calibration'.format(CUR_LANE.name.upper()))
-                return True, ""
-
-            self.AFC.gcode.respond_info('Calibrating {}'.format(CUR_LANE.name.upper()))
-            # reset to extruder
-            calc_position(CUR_LANE, lambda: CUR_LANE.load_state, 0, short_dis, tol)
-            hub_pos = calibrate_hub(CUR_LANE, CUR_HUB)
-            if CUR_HUB.state:
-                CUR_LANE.move(CUR_HUB.move_dis * -1, self.AFC.short_moves_speed, self.AFC.short_moves_accel, True)
-            CUR_LANE.loaded_to_hub = True
-            CUR_LANE.do_enable(False)
-            cal_msg = "\n{} dist_hub: {}".format(CUR_LANE.name.upper(), (hub_pos - CUR_HUB.hub_clear_move_dis))
-            return True, cal_msg
-
-        # Determine if a specific lane is provided
-        if lanes is not None:
-            self.AFC.gcode.respond_info('Starting AFC distance Calibrations')
-            cal_msg += 'AFC Calibration distances +/-{}mm'.format(tol)
-            cal_msg += '\n<span class=info--text>Update values in AFC_Hardware.cfg</span>'
-            if lanes != 'all':
-                lane_to_calibrate = find_lane_to_calibrate(lanes)
-                if lane_to_calibrate is None:
-                    return
-                # Calibrate the specific lane
-                checked, msg = calibrate_lane(lane_to_calibrate)
-                if(not checked): return
-                cal_msg += msg
-            else:
-                # Calibrate all lanes if no specific lane is provided
-                for UNIT in self.AFC.units.keys():
-                    for LANE in self.AFC.units[UNIT].keys():
-                        # Calibrate the specific lane
-                        checked, msg = calibrate_lane(LANE)
-                        if(not checked): return
-                        cal_msg += msg
-
-        else:
-            cal_msg +='No lanes selected to calibrate dist_hub'
-
-        if afc_bl is not None:
-            if lanes is None:
-                self.AFC.gcode.respond_info('Starting AFC distance Calibrations')
-                cal_msg += 'AFC Calibration distances +/-{}mm'.format(tol)
-                cal_msg += '\n<span class=info--text>Update values in AFC_Hardware.cfg</span>'
-
-            lane_to_calibrate = find_lane_to_calibrate(afc_bl)
-
-            if lane_to_calibrate is None:
-                return
-
-            lane = lane_to_calibrate
-            CUR_LANE = self.printer.lookup_object('AFC_stepper ' + lane)
-            CUR_EXTRUDER = self.printer.lookup_object('AFC_extruder ' + CUR_LANE.extruder_name)
-            CUR_HUB = self.printer.lookup_object('AFC_hub ' + CUR_LANE.unit)
-            self.AFC.gcode.respond_info('Calibrating Bowden Length with {}'.format(CUR_LANE.name.upper()))
-
-            move_until_state(CUR_LANE, lambda: CUR_HUB.state, CUR_HUB.move_dis, tol, short_dis)
-
-            bow_pos = 0
-            if CUR_EXTRUDER.tool_start:
-                while not CUR_EXTRUDER.tool_start_state:
-                    CUR_LANE.move(dis, self.AFC.short_moves_speed, self.AFC.short_moves_accel)
-                    bow_pos += dis
-                    self.AFC.reactor.pause(self.AFC.reactor.monotonic() + 0.1)
-                bow_pos = calc_position(CUR_LANE, lambda: CUR_EXTRUDER.tool_start_state, bow_pos, short_dis, tol)
-                CUR_LANE.move(bow_pos * -1, self.AFC.long_moves_speed, self.AFC.long_moves_accel, True)
-                calibrate_hub(CUR_LANE, CUR_HUB)
-                if CUR_HUB.state:
-                    CUR_LANE.move(CUR_HUB.move_dis * -1, self.AFC.short_moves_speed, self.AFC.short_moves_accel, True)
-                if CUR_EXTRUDER.tool_start == 'buffer':
-                    cal_msg += '\n afc_bowden_length: {}'.format(bow_pos - (short_dis * 2))
-                else:
-                    cal_msg += '\n afc_bowden_length: {}'.format(bow_pos - short_dis)
-                CUR_LANE.do_enable(False)
-            else:
-                self.AFC.gcode.respond_info('CALIBRATE_AFC is not currently supported without tool start sensor')
-
-        self.AFC.save_vars()
-        self.AFC.gcode.respond_info(cal_msg)
-
-def load_config(config):
+def load_config_prefix(config):
     return afcBoxTurtle(config)
