--- extras/AFC_screen.py
@@ -1,14 +1,0 @@
-# Armored Turtle Automated Filament Changer
-#
-# Copyright (C) 2024 Armored Turtle
-#
-# This file may be distributed under the terms of the GNU GPLv3 license.
-
-class AFCscreen:
-    def __init__(self, config):
-        self.printer = config.get_printer()
-        self.name = config.get_name().split()[-1]
-        self.mac = config.get('mac_address', None)
-
-def load_config_prefix(config):
-    return AFCscreen(config)
